0001641: PD Spikes
git-svn-id: https://mabako.net:8081@566 8706e3ef-0d6a-4709-9ab6-f16f3f3d4656

+++ src_trunk/resources/spike-system/s_spikes.lua
@@ -1,7 +1,9 @@
 TotalSpikes=nil
 Spike = {}
 SpikeLimit=10
-Shape= {}
+Shape1= {}
+Shape2= {}
+
 function PlacingSpikes(sourcePlayer, command)
 	local theTeam = getPlayerTeam(sourcePlayer)
 	local teamType = getElementData(theTeam, "type")
@@ -35,6 +37,65 @@ function PlacingSpikes(sourcePlayer, command)
 end
 addCommandHandler("deployspikes", PlacingSpikes)
 
+function ThrowSpikes(sourcePlayer, command)
+	local theTeam = getPlayerTeam(sourcePlayer)
+	local teamType = getElementData(theTeam, "type")
+	
+	--if (teamType==2) then
+		local x1,y1,z1 = getElementPosition(sourcePlayer)		
+		local rotz = getPedRotation(sourcePlayer)
+			if(TotalSpikes == nil or TotalSpikes < SpikeLimit) then
+			if(TotalSpikes == nil) then
+				TotalSpikes = 1
+			else
+				TotalSpikes = TotalSpikes+1
+			end
+			for value=1,SpikeLimit,1 do
+				if(Spike[value] == nil) then
+					-- some general stuff
+					local px, py, pz = getElementPosition ( sourcePlayer )
+					local rz = getPedRotation ( sourcePlayer )  
+					
+					
+					-- some calculations to find the place for the object
+					local distance = 5
+					local x = distance*math.cos((rz+90)*math.pi/180)
+					local y = distance*math.sin((rz+90)*math.pi/180)
+					local b2 = 15 / math.cos(math.pi/180)
+					local nx = px + x
+					local ny = py + y
+					local nz = pz - 0.8
+
+					Spike[value] = createObject ( 2892, nx, ny, nz, 0.0, 0.0, rz)
+					exports.pool:allocateElement(Spike[value])
+					
+					
+					-- Object is done, now we need the colpolygen
+					-- size of the object:
+					--	1		10.0 		3
+					--  x-------------------x
+					--  |                   | 1.0  
+					--	x-------------------x
+					--  2					4
+
+					
+					-- create the colpolygon
+					Shape1[value] = createColRectangle( (nx - 5), (ny - 5), 5.0, 5.0 )
+					exports.pool:allocateElement(Shape1[value])
+					Shape2[value] = createColRectangle( (nx), (ny), 5.0, 5.0 )
+					exports.pool:allocateElement(Shape2[value])
+					setElementData(Shape1[value], "type", "spikes")
+					setElementData(Shape2[value], "type", "spikes")
+					outputChatBox("Spawned spikes with ID:" .. value, sourcePlayer, 0, 194, 0)
+					break
+				end
+			end
+		else
+			outputChatBox("Too many spikes are already spawned.", sourcePlayer, 255, 194, 14)
+		end
+	--end
+end
+addCommandHandler("throwspikes", ThrowSpikes)
 
 function RemovingSpikes(sourcePlayer, command, ID)
 	local theTeam = getPlayerTeam(sourcePlayer)
@@ -55,8 +116,12 @@ function RemovingSpikes(sourcePlayer, command, ID)
 					TotalSpikes = TotalSpikes -1
 					destroyElement(Spike[message])
 					Spike[message] = nil
-					destroyElement(Shape[message])
-					Shape[message] = nil
+					destroyElement(Shape1[message])
+					Shape1[message] = nil
+					if (isElement(Shape2[message]) then
+						destroyElement(Shape2[message])
+						Shape2[message] = nil
+					end
 					if(TotalSpikes <= 0) then
 						TotalSpikes = nil
 					end
@@ -72,17 +137,23 @@ end
 addCommandHandler("removespikes", RemovingSpikes)
 
 function AdminRemovingSpikes(sourcePlayer, command)
-	for value=1,SpikeLimit,1 do
-		if(Spike[value] ~= nil) then
-			local id = tonumber ( value )
-			destroyElement(Spike[id])
-			Spike[id] = nil
-			destroyElement(Shape[id])
-			Shape[id] = nil
+	if	(exports.global:isPlayerAdmin(sourcePlayer)) then
+		for value=1,SpikeLimit,1 do
+			if(Spike[value] ~= nil) then
+				local id = tonumber ( value )
+				destroyElement(Spike[id])
+				Spike[id] = nil
+				destroyElement(Shape1[id])
+				Shape1[id] = nil
+				if (isElement(Shape2[id]) then
+					destroyElement(Shape2[id])
+					Shape2[id] = nil
+				end
+			end
 		end
+		outputChatBox("Removed all the spawned spikes.", sourcePlayer, 0, 194, 0)
+		TotalSpikes = nil
 	end
-	outputChatBox("Removed all the spawned spikes.", sourcePlayer, 0, 194, 0)
-	TotalSpikes = nil
 end
 addCommandHandler("aremovespikes", AdminRemovingSpikes)
